Refactor internals.
- Merge the apply* set of helpers into applyTimestamp() and applyPayload() respectively. Both get inlined, albeit applyTimestamp() needs to keep the "& 1" masking outside of the func to not exceed the inlining cost, leading to a tiny bit of duplication;
- New applyPayload() also falls back to std lib for the big endian store to make use of the compiler optimizations. This shaves off another about 0.7ns per generation call;
- Partition is now stored as a big endian shifted uint32 internally, instead of the struct. This allows it to get simply bitwise ORed with the sequence during generation, shaving off some instructions. The actual size of the Generator remains unchanged - it does not grow by 2 bytes due to the memory alignment of the Generator's fields. The public representation of Partitions remains unchanged;
- Times are now handled as unsigned internally;

+++ time_test.go
@@ -7,7 +7,7 @@ import (
 
 func TestTime_Nanotime(t *testing.T) {
 	actual := nanotime()
-	expected := (time.Now().UnixNano() - epochNsec) / TimeUnit
+	expected := uint64(time.Now().UnixNano()-epochNsec) / TimeUnit
 
 	if actual != expected {
 		t.Errorf("expected [%v], got [%v]", expected, actual)
